Concepts rename pt2
* renamed stub files and adjusted config file.  Corrected reference to floating point numbers and added additonal info on precision.

* renamed stub files and corrected config file.

* Edited code commet formatting and adjusted content.

* Added missing uuids and changed concept names for strings.

* Changed the name of str related concepts and numbers related concepts.  Updated numbers exercise info.

* Added additional links.  Renamed a few concepts.

* renamed stub files and adjusted config file.  Corrected reference to floating point numbers and added additonal info on precision.

* renamed stub files and corrected config file.

* Edited code commet formatting and adjusted content.

* Added missing uuids and changed concept names for strings.

* Changed the name of str related concepts and numbers related concepts.  Updated numbers exercise info.

* Added additional links.  Renamed a few concepts.

* ran prettier. Again.

* Ran prettier for currency exchange.

* Updated concept WIP table with numbers exercise and reanmed heirarchical concepts.

* Update languages/exercises/README.md


* Update languages/exercises/README.md
